pyproject.toml
[project]
name = "excelipy"
version = "0.1.26"
description = "Wrapper around xlsxwriter to improve usability"
readme = "README.md"
requires-python = ">=3.9.10"

dependencies = [
    "numpy>=2.0.2",
    "pandas>=2.2.3",
    "pillow>=11.3.0",
    "pydantic>=2.11.3",
    "xlsxwriter>=3.2.2",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.3.3",
    "pytest-cov>=4.1.0",
    "coverage>=7.6.1",
    "pytest-mock>=3.14.0",
    "pytest-asyncio>=0.25.3",
    "debugpy>=1.8.13",
    "pyright>=1.1.398",
]

[tool.setuptools]
packages = ["excelipy"]
